requirements.txt
absl-py==0.11.0
apturl==0.5.2
asn1crypto==0.24.0
Brlapi==0.6.6
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
et-xmlfile==1.0.1
Glymur==0.9.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
httplib2==0.9.2
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
jdcal==1.4.1
joblib==1.0.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lmdb==1.0.0
louis==3.5.0
lxml==4.6.2
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.3
MarkupSafe==1.0
matplotlib==3.3.3
netifaces==0.10.4
networkx==2.5
numpy==1.19.5
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
opencv-python==4.4.0.46
openpyxl==3.0.6
openslide-python==1.1.2
pandas==1.1.5
pexpect==4.2.1
Pillow==8.1.0
protobuf==3.14.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.4.1
reportlab==3.4.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.7
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
screen-resolution-extra==0.0.0
seaborn==0.11.1
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
threadpoolctl==2.1.0
tiatoolbox==0.4.0
tifffile==2020.9.3
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
tqdm==4.56.0
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==1.0.1
xkit==0.0.0
xlrd==1.2.0
zipp==3.4.0
zope.interface==4.3.2
